fixes[t:535]

+++ engine/core/controller/enginesettings.cpp
@@ -76,9 +76,9 @@ namespace FIFE {
 	EngineSettings::~EngineSettings() {
 	}
 
-	void EngineSettings::setBitsPerPixel(uint16_t bitsperpixel) {
-		std::vector<uint16_t> pv = getPossibleBitsPerPixel();
-		std::vector<uint16_t>::iterator i = std::find(pv.begin(), pv.end(), bitsperpixel);
+	void EngineSettings::setBitsPerPixel(uint8_t bitsperpixel) {
+		std::vector<uint8_t> pv = getPossibleBitsPerPixel();
+		std::vector<uint8_t>::iterator i = std::find(pv.begin(), pv.end(), bitsperpixel);
 		if (i != pv.end()) {
 			m_bitsperpixel = bitsperpixel;
 			return;
@@ -91,8 +91,8 @@ namespace FIFE {
 		m_bitsperpixel = 0;  //default value
 	}
 
-	std::vector<uint16_t> EngineSettings::getPossibleBitsPerPixel() const {
-		std::vector<uint16_t> tmp;
+	std::vector<uint8_t> EngineSettings::getPossibleBitsPerPixel() const {
+		std::vector<uint8_t> tmp;
 		tmp.push_back(0);
 		tmp.push_back(16);
 		tmp.push_back(24);
