+++ engine/core/model/metamodel/grids/squaregrid.h
@@ -43,7 +43,7 @@ namespace FIFE {
 		const std::string& getType() const;
 		const std::string& getName() const;
 		bool isAccessible(const ModelCoordinate& curpos, const ModelCoordinate& target);
-		float getAdjacentCost(const ModelCoordinate& curpos, const ModelCoordinate& target);
+		double getAdjacentCost(const ModelCoordinate& curpos, const ModelCoordinate& target);
 		uint32_t getCellSideCount() const { return 4; }
 		ExactModelCoordinate toMapCoordinates(const ExactModelCoordinate& layer_coords);
 		ModelCoordinate toLayerCoordinates(const ExactModelCoordinate& map_coord);
